v0.0.0.5
Fixed the MAL search-bar in the lightbox when "Big search results"
option is turned on.
Previously I added style modifications code to make the search-bar
bigger together with the search results' images, now MAL has changed
their design and made the search-bar big by default, and also much
better looking (finally!). because of that change, my modifications to
the search-bar are breaking the new design and are unnecessary anymore,
therefore I removed my style modifications of the search-bar for it to
stay as is.